Create tests for track, playlist and unkown entity
